+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - More documentation: a Jupyter interactive notebook of Plotters tutorial 
 
+### Improvement
+
+- More style conversion APIs
+
 ## Plotters 0.2.1 (2019-06-10)
 
 ### Improvement